avoid making a bad symlink
Avoid making a bad symlink to avoid trouble with the nightly server.

+++ nightly.sh
@@ -11,4 +11,8 @@ git submodule update --init
 cargo bench
 
 # make report/index.html accessible in the criterion/ directory
-ln -s target/criterion/report/index.html target/criterion/
+if [ -e target/criterion/report/index.html ]; then
+    ln -s target/criterion/report/index.html target/criterion/
+else
+    echo "<html><body>Hey! Something went wrong and <code>target/criterion/report/index.html</code> doesn't exist!</body></html>" > target/criterion/index.html
+fi
